requirements.txt
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
